+++ changelog.md
@@ -1,2 +1,3 @@
 - Added demo command
 - Fixed tank rendering
+- Added a car pressure plate